+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Current (in progress)
 
 - Use beta admin as default [#621](https://github.com/datagouv/udata-front/pull/621)
+- Fix following cache issue [#631](https://github.com/datagouv/udata-front/pull/631)
+- Add dataservice search to header [#622](https://github.com/datagouv/udata-front/pull/622)
+
+## 6.0.7 (2024-12-20)
+
+- Add NATIONAL_MOURNING config [#625](https://github.com/datagouv/udata-front/pull/625)
 
 ## 6.0.6 (2024-12-09)